refactor: 俺好みにリファクタしたよ
