+++ web/src/lib/r2/s3.ts
@@ -1,0 +1,10 @@
+import { S3Client } from "@aws-sdk/client-s3";
+
+export const client = new S3Client({
+  region: "auto",
+  endpoint: process.env.R2_ENDPOINT_URL || "",
+  credentials: {
+    accessKeyId: process.env.R2_ACCESS_KEY_ID || "",
+    secretAccessKey: process.env.R2_SECRET_ACCESS_KEY || "",
+  },
+});
